@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -22,9 +21,8 @@ ActiveRecord::Schema.define(version: 20151215123545) do
     t.datetime "created_at",                      null: false
     t.datetime "updated_at",                      null: false
     t.boolean  "dish_of_the_day", default: false
+    t.index ["dish_id"], name: "index_dish_days_on_dish_id", using: :btree
   end
-
-  add_index "dish_days", ["dish_id"], name: "index_dish_days_on_dish_id", using: :btree
 
   create_table "dishes", force: :cascade do |t|
     t.string   "name"
